Updates to recent DOLFIN

+++ dolfintape/MatrixView.h
@@ -92,7 +92,7 @@ namespace dolfin
     }
 
     /// Return local ownership range
-    virtual std::pair<std::size_t, std::size_t>
+    virtual std::pair<std::int64_t, std::int64_t>
       local_range(std::size_t dim) const
     {
       dolfin_not_implemented();
